FoodSource Burger

+++ app/src/main/java/com/example/cooked_fever/appliances/Appliance.java
@@ -8,6 +8,6 @@ public interface Appliance {
     void draw(Canvas canvas);
 
     boolean isReady();
-    boolean onClick(float x, float y);  // return true if this appliance was interacted with
+    boolean onClick(int x, int y);  // return true if this appliance was interacted with
     Rect getHitbox();
 }
