+++ app/src/main/java/com/example/cooked_fever/appliances/CocaColaMaker.java
@@ -1,9 +1,6 @@
 package com.example.cooked_fever.appliances;
 
-import android.graphics.Canvas;
-import android.graphics.Color;
-import android.graphics.Paint;
-import android.graphics.Rect;
+import android.graphics.*;
 
 import com.example.cooked_fever.appliances.Appliance;
 
@@ -115,8 +112,8 @@ public class CocaColaMaker implements Appliance {
     }
 
     @Override
-    public boolean onClick(float x, float y) {
-        if (hitbox.contains((int) x, (int) y)) {
+    public boolean onClick(int x, int y) {
+        if (hitbox.contains(x, y)) {
             takeGlass();
             return true;
         }
